Fix hardest send grade ordering
